Plugin/VideoAnalyzer/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
yt-dlp>=2023.0.0

